DocId not present for skip patient fixed
